refactor(frontend): extract save logic for project and user settings into use cases
